Move approximation and gaussian files to utils
